mqtt_listener/requirements.txt
paho-mqtt==2.1.0
python-dotenv==0.21.0
psycopg2-binary==2.9.10
sqlmodel==0.0.24
sqlalchemy==2.0.43
alembic==1.13.1